Remove redundant code from `PacketLib1124` Based on https://github.com/Dawn-of-Light/DOLSharp/commit/5f83d20ba0dae6fd855daef1ccf76e28b80afa61 Attempted to keep Atlas changes where it made sense, but some old clients aren't guaranteed to work properly since legacy libs weren't kept updated (for example there are still references to `GameLiving.EffectList`). General cleanup.

+++ GameServer/packets/Server/PacketLib169.cs
@@ -1,28 +1,24 @@
 /*
  * DAWN OF LIGHT - The first free open source DAoC server emulator
- * 
+ *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License
  * as published by the Free Software Foundation; either version 2
  * of the License, or (at your option) any later version.
- * 
+ *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU General Public License
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
  *
  */
 
-using System;
-using System.Collections;
-using DOL.Database;
-using DOL.GS.Spells;
-using DOL.GS.Styles;
 using DOL.GS.Effects;
+using DOL.GS.Spells;
 
 namespace DOL.GS.PacketHandler
 {
@@ -39,12 +35,13 @@ namespace DOL.GS.PacketHandler
 
 		public override void SendGroupWindowUpdate()
 		{
-			if (m_gameClient.Player == null) return;
+			if (m_gameClient.Player == null)
+				return;
 
 			using (GSTCPPacketOut pak = new GSTCPPacketOut(GetPacketCode(eServerPackets.VariousUpdate)))
 			{
 				pak.WriteByte(0x06);
-	
+
 				Group group = m_gameClient.Player.Group;
 				if (group == null)
 				{
@@ -52,25 +49,25 @@ namespace DOL.GS.PacketHandler
 				}
 				else
 				{
-					pak.WriteByte((byte)group.MemberCount);
+					pak.WriteByte(group.MemberCount);
 				}
-	
+
 				pak.WriteByte(0x01);
 				pak.WriteByte(0x00);
-	
+
 				if (group != null)
 				{
 					foreach (GameLiving living in group.GetMembersInTheGroup())
 					{
 						bool sameRegion = living.CurrentRegion == m_gameClient.Player.CurrentRegion;
-	
+
 						pak.WriteByte(living.Level);
 						if (sameRegion)
 						{
 							pak.WriteByte(living.HealthPercentGroupWindow);
 							pak.WriteByte(living.ManaPercent);
 							pak.WriteByte(living.EndurancePercent); //new in 1.69
-	
+
 							byte playerStatus = 0;
 							if (!living.IsAlive)
 								playerStatus |= 0x01;
@@ -84,11 +81,11 @@ namespace DOL.GS.PacketHandler
 								playerStatus |= 0x10;
 							if (living.CurrentRegion != m_gameClient.Player.CurrentRegion)
 								playerStatus |= 0x20;
-	
+
 							pak.WriteByte(playerStatus);
-							// 0x00 = Normal , 0x01 = Dead , 0x02 = Mezzed , 0x04 = Diseased , 
+							// 0x00 = Normal , 0x01 = Dead , 0x02 = Mezzed , 0x04 = Diseased ,
 							// 0x08 = Poisoned , 0x10 = Link Dead , 0x20 = In Another Region
-	
+
 							pak.WriteShort((ushort)living.ObjectID);//or session id?
 						}
 						else
@@ -136,7 +133,7 @@ namespace DOL.GS.PacketHandler
 					playerStatus |= 0x20;
 
 				pak.WriteByte(playerStatus);
-				// 0x00 = Normal , 0x01 = Dead , 0x02 = Mezzed , 0x04 = Diseased , 
+				// 0x00 = Normal , 0x01 = Dead , 0x02 = Mezzed , 0x04 = Diseased ,
 				// 0x08 = Poisoned , 0x10 = Link Dead , 0x20 = In Another Region
 
 				if (updateIcons)
